get service area emails gateway
